// projects/client/src/app/image-gallery/image-gallery.conf.ts
// tslint:disable-next-line:class-name
export interface GALLERY_IMAGE {
  _cached?: boolean;
  _id?: string;
  url: string;
  thumbnailUrl?: string;
  altText?: string;
  title?: string;
  extUrl?: string;
  extUrlTarget?: string;
  productUrl?: string;
  productTitle?: string;
  galleryTitle?: string;
  productImageUrl?: string;
}